[core-v2.4.0-alpha.6, react-v2.2.3-alpha.4, vue v2.1.3-alpha.4] : Update - Refactor updateBalance within notification trigger (#1129)
* Refactor updateBalance within notification trigger
* Rename update-balances file to conform to kebab case
