+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@web3-onboard/core",
-  "version": "2.4.0-alpha.5",
+  "version": "2.4.0-alpha.6",
   "repository": "blocknative/web3-onboard",
   "scripts": {
     "build": "rollup -c",
