+++ packages/demo/package.json
@@ -23,7 +23,7 @@
   },
   "dependencies": {
     "@web3-onboard/coinbase": "^2.0.7",
-    "@web3-onboard/core": "^2.3.2",
+    "@web3-onboard/core": "^2.4.0-alpha.6",
     "@web3-onboard/dcent": "^2.0.4",
     "@web3-onboard/fortmatic": "^2.0.6",
     "@web3-onboard/gnosis": "^2.0.5",
