+++ packages/react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@web3-onboard/react",
-  "version": "2.2.3-alpha.3",
+  "version": "2.2.3-alpha.4",
   "description": "Collection of React Hooks for web3-onboard",
   "repository": "blocknative/web3-onboard",
   "module": "dist/index.js",
@@ -24,7 +24,7 @@
     "typescript": "^4.5.5"
   },
   "dependencies": {
-    "@web3-onboard/core": "^2.4.0-alpha.5",
+    "@web3-onboard/core": "^2.4.0-alpha.6",
     "@web3-onboard/common": "^2.1.4",
     "use-sync-external-store": "1.0.0"
   },
